Actualización de Header.h y optimización de tipos de variables. Eliminación de 'cout', uso de memoria dinámica y mejoras en las transformaciones

+++ Desafio1/Header.h
@@ -1,9 +1,7 @@
-#include <fstream>
-#include <iostream>
 #include <QCoreApplication>
 #include <QImage>
-#include <bitset>
 #include <QDir>
+#include <QString>
 
 using namespace std;
 
@@ -11,12 +9,12 @@ using namespace std;
 #define HEADER_H
 unsigned char* loadPixels(QString input, int &width, int &height);
 bool exportImage(unsigned char* pixelData, int width,int height, QString archivoSalida);
-unsigned char *funcionXOR(unsigned char* pixelData, unsigned char* pixelImMascara, int dataSize, int n);
-unsigned char* funcionOR(unsigned char* pixelData, unsigned char* pixelImMascara, int dataSize, int n);
-unsigned char* funcionAND(unsigned char* pixelData, unsigned char* pixelImMascara, int dataSize, int n);
+unsigned char *funcionXOR(unsigned char* pixelData, unsigned char* pixelImMascara, size_t dataSize, uint8_t n);
+unsigned char* funcionOR(unsigned char* pixelData, unsigned char* pixelImMascara, size_t dataSize, uint8_t n);
+unsigned char* funcionAND(unsigned char* pixelData, unsigned char* pixelImMascara, size_t dataSize, uint8_t n);
 unsigned int* loadSeedMasking(const char* nombreArchivo, int &seed, int &n_pixels);
 bool verificationTransformation(unsigned int* RGB , unsigned char* mascara, unsigned char* transformation, int &seed, int &n_pixels);
-unsigned char * funcionRR(unsigned char *pixelData, unsigned char* pixelImMascara, int dataSize, int n);
-unsigned char * funcionRL (unsigned char *pixelData, unsigned char* pixelImMascara, int dataSize, int n);
+unsigned char * funcionRR(unsigned char *pixelData, unsigned char* pixelImMascara, size_t dataSize, uint8_t n);
+unsigned char * funcionRL (unsigned char *pixelData, unsigned char* pixelImMascara, size_t dataSize, uint8_t n);
 
 #endif // HEADER_H
